refactor: settings tab and category manager modal logic extraction, sub-components division (#305)
